refactor(agents): :recycle: move common fixtures of test into conftest
